Changement csv et json terminé

+++ kernel/src/main/java/dataextraction/Extractor.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 public interface Extractor {
 
     public Object extractNextValue(String s);
-    public int getCurrentTime();
+    public long getCurrentTime();
     public boolean isFinish();
 
 }
